refactor: extract chat input helpers and AttachedReviewsPanel
Decompose ChatInput/index.tsx by extracting:

- Skill invocation helpers (parseCommandWithSkillInvocation, buildSkillInvocationMetadata, etc.)
- Runtime validation (validateCreationRuntime)
- Attachment conversion (filePartsToChatAttachments)
- Review data helpers (getReviewData, getReviewIds)

Also extract AttachedReviewsPanel component for rendering attached reviews.

index.tsx: 2637 → 2443 lines (-194 lines)
